(refactor) Move user routes into route and controller files
