+++ server/barUsers/barUserRoutes.js
@@ -1,0 +1,5 @@
+var barUserController = require('./barUserController');
+
+module.exports = function(app){
+  
+};
