+++ server/users/userRoutes.js
@@ -1,0 +1,6 @@
+var userController = require('./userController.js');
+
+module.exports = function(app){
+  app.post('/login', userController.login);
+  app.post('/signup', userController.signup);
+};
